chore: move change notifier into dedicated directory
