+++ webapp/frontend/lib/change_notifier/quote_change_notifier.dart
