+++ webapp/frontend/lib/trade/position_table.dart
@@ -1,16 +1,16 @@
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
+import 'package:get_10101/change_notifier/quote_change_notifier.dart';
 import 'package:get_10101/common/amount_text.dart';
 import 'package:get_10101/common/color.dart';
-import 'package:get_10101/common/currency_change_notifier.dart';
+import 'package:get_10101/change_notifier/currency_change_notifier.dart';
 import 'package:get_10101/common/direction.dart';
 import 'package:get_10101/common/model.dart';
-import 'package:get_10101/settings/channel_change_notifier.dart';
+import 'package:get_10101/change_notifier/channel_change_notifier.dart';
 import 'package:get_10101/settings/dlc_channel.dart';
 import 'package:get_10101/trade/close_position_confirmation_dialog.dart';
-import 'package:get_10101/trade/position_change_notifier.dart';
+import 'package:get_10101/change_notifier/position_change_notifier.dart';
 import 'package:get_10101/services/position_service.dart';
-import 'package:get_10101/trade/quote_change_notifier.dart';
 import 'package:get_10101/services/quote_service.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
