+++ webapp/frontend/lib/wallet/history_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get_10101/wallet/onchain_payment_history_item.dart';
-import 'package:get_10101/wallet/wallet_change_notifier.dart';
+import 'package:get_10101/change_notifier/wallet_change_notifier.dart';
 import 'package:get_10101/services/wallet_service.dart';
 import 'package:provider/provider.dart';
